+++ src/Snow.Blog/Snow.Blog.Repository/BaseRepository.cs
@@ -360,14 +360,16 @@ namespace Snow.Blog.Repository
         }
 
         /// <summary>
-        /// 批量插入 Entity
+        /// 插入
         /// </summary>
-        /// <param name="entity">实体集合</param>
+        /// <param name="entity">实体</param>
+        /// <param name="connection">数据库连接</param>
         /// <param name="transaction">事物</param>
         /// <param name="commandTimeout">超时时间</param>
-        /// <returns>受影响的行数</returns>
-        public async Task<long> InsertAsync(TEntity entity, IDbConnection connection = null, IDbTransaction transaction = null, int? commandTimeout = null)
+        /// <returns>主键</returns>
+        public async Task<int> InsertAsync(TEntity entity, IDbConnection connection = null, IDbTransaction transaction = null, int? commandTimeout = null)
         {
+            // TODO:测试Dapper.Contrib各Insert方法
             if (connection == null)
             {
                 using (connection = DBSessionFactory.CreateDbConnection(ConnectionString))
